# Speedup compilation of index vectors