Refactor `colptrs` and `rowvals` functions for improved compile time and partially runtime performance:
* avoid switch statements with potentially thousands of cases
* avoid element-wise assignments for large arrays where all values are assigned
* instead, keep relevant data in static (arrays of) arrays and just copy to destination

Tested with `dJydy_colptrs.cpp` 19.1MB 592139 lines (ny = 768):
* file size old: 19.1MB; file size new: 1.7MB  (**10x!**)
* -O2:
  * RAM: old: 1.6GB, new: 146MB (**10x!**)
  * User time: old: 5min9s, new: 1.03s (**300x!**)
  * run time performance (1e4 times retrieving all dJydy_colptrs): old: 11.516s, new: 5.938s (**2x**)
* -O3: 
  * RAM: old: 15.7GB, new: 146MB (**100x!**)
  * User time: old: 6min12s, new: 1.07s (**300x!**)
  * run time performance (1e4 times retrieving all dJydy_colptrs): old: 5.696s, new: 5.356s